Extract block view holders into separate classes for better readibility
