+++ WordPress/src/main/java/org/wordpress/android/ui/stats/refresh/lists/sections/BlockListAdapter.kt
@@ -39,22 +39,23 @@ import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItem.Type.
 import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItem.Type.VALUE_ITEM
 import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItem.Type.values
 import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItem.ValueItem
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.BarChartViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.ChartLegendViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.DividerViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.EmptyViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.ExpandableItemViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.FourColumnsViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.HeaderViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.InformationViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.LinkViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.ListItemViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.ListItemWithIconViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.MapViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.TabsViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.TextViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.TitleViewHolder
-import org.wordpress.android.ui.stats.refresh.lists.sections.BlockListItemViewHolder.ValueViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.BlockListItemViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.BarChartViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.ChartLegendViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.DividerViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.EmptyViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.ExpandableItemViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.FourColumnsViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.HeaderViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.InformationViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.LinkViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.ListItemViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.ListItemWithIconViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.MapViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.TabsViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.TextViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.TitleViewHolder
+import org.wordpress.android.ui.stats.refresh.lists.sections.viewholders.ValueViewHolder
 import org.wordpress.android.util.image.ImageManager
 
 class BlockListAdapter(val imageManager: ImageManager) : Adapter<BlockListItemViewHolder>() {
@@ -89,18 +90,29 @@ class BlockListAdapter(val imageManager: ImageManager) : Adapter<BlockListItemVi
     override fun onCreateViewHolder(parent: ViewGroup, itemType: Int): BlockListItemViewHolder {
         return when (values()[itemType]) {
             TITLE -> TitleViewHolder(parent)
-            LIST_ITEM_WITH_ICON -> ListItemWithIconViewHolder(parent, imageManager)
+            LIST_ITEM_WITH_ICON -> ListItemWithIconViewHolder(
+                    parent,
+                    imageManager
+            )
             LIST_ITEM -> ListItemViewHolder(parent)
             EMPTY -> EmptyViewHolder(parent)
             TEXT -> TextViewHolder(parent)
             COLUMNS -> FourColumnsViewHolder(parent)
             LINK -> LinkViewHolder(parent)
             BAR_CHART -> BarChartViewHolder(parent)
-            CHART_LEGEND -> ChartLegendViewHolder(parent)
-            TABS -> TabsViewHolder(parent, imageManager)
+            CHART_LEGEND -> ChartLegendViewHolder(
+                    parent
+            )
+            TABS -> TabsViewHolder(
+                    parent,
+                    imageManager
+            )
             INFO -> InformationViewHolder(parent)
             HEADER -> HeaderViewHolder(parent)
-            EXPANDABLE_ITEM -> ExpandableItemViewHolder(parent, imageManager)
+            EXPANDABLE_ITEM -> ExpandableItemViewHolder(
+                    parent,
+                    imageManager
+            )
             DIVIDER -> DividerViewHolder(parent)
             MAP -> MapViewHolder(parent)
             VALUE_ITEM -> ValueViewHolder(parent)
