+++ WordPress/src/main/java/org/wordpress/android/ui/stats/refresh/lists/sections/viewholders/BlockListItemViewHolder.kt
@@ -1,0 +1,11 @@
+package org.wordpress.android.ui.stats.refresh.lists.sections.viewholders
+
+import android.support.annotation.LayoutRes
+import android.support.v7.widget.RecyclerView.ViewHolder
+import android.view.LayoutInflater
+import android.view.ViewGroup
+
+open class BlockListItemViewHolder(
+    parent: ViewGroup,
+    @LayoutRes layout: Int
+) : ViewHolder(LayoutInflater.from(parent.context).inflate(layout, parent, false))
